requirements.txt
discord.py>=1.7.1
discord-pretty-help>=1.3.3
python-Levenshtein-wheels>=0.13.1
aiohttp>=3.7.4
asyncpg>=0.20.1
python-dotenv>=0.13.0
yoyo-migrations>=7.0.2
psycopg2-binary>=2.8.5
yarl==1.4.2
steam>=1.2.0